@@ -3,7 +3,7 @@ package aristoteles
 import (
 	"bytes"
 	"fmt"
-	"github.com/elastic/go-elasticsearch/v8"
+	"github.com/elastic/go-elasticsearch/v9"
 	"io"
 	"net/http"
 	"os"
