@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
 // 	protoc-gen-go v1.28.1
-// 	protoc        v3.21.12
+// 	protoc        v6.33.2
 // source: proto/eupalinos.proto
 
 package proto
@@ -72,10 +72,14 @@ type InternalEpistello struct {
 	sizeCache     protoimpl.SizeCache
 	unknownFields protoimpl.UnknownFields
 
-	Id      string `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`
-	Data    string `protobuf:"bytes,2,opt,name=data,proto3" json:"data,omitempty"`
-	Channel string `protobuf:"bytes,4,opt,name=channel,proto3" json:"channel,omitempty"`
-	Traceid string `protobuf:"bytes,5,opt,name=traceid,proto3" json:"traceid,omitempty"`
+	Id string `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`
+	// Types that are assignable to Payload:
+	//
+	//	*InternalEpistello_Data
+	//	*InternalEpistello_BytesData
+	Payload isInternalEpistello_Payload `protobuf_oneof:"payload"`
+	Channel string                      `protobuf:"bytes,4,opt,name=channel,proto3" json:"channel,omitempty"`
+	Traceid string                      `protobuf:"bytes,5,opt,name=traceid,proto3" json:"traceid,omitempty"`
 }
 
 func (x *InternalEpistello) Reset() {
@@ -117,11 +121,25 @@ func (x *InternalEpistello) GetId() string {
 	return ""
 }
 
+func (m *InternalEpistello) GetPayload() isInternalEpistello_Payload {
+	if m != nil {
+		return m.Payload
+	}
+	return nil
+}
+
 func (x *InternalEpistello) GetData() string {
-	if x != nil {
+	if x, ok := x.GetPayload().(*InternalEpistello_Data); ok {
 		return x.Data
 	}
 	return ""
+}
+
+func (x *InternalEpistello) GetBytesData() []byte {
+	if x, ok := x.GetPayload().(*InternalEpistello_BytesData); ok {
+		return x.BytesData
+	}
+	return nil
 }
 
 func (x *InternalEpistello) GetChannel() string {
@@ -137,6 +155,22 @@ func (x *InternalEpistello) GetTraceid() string {
 	}
 	return ""
 }
+
+type isInternalEpistello_Payload interface {
+	isInternalEpistello_Payload()
+}
+
+type InternalEpistello_Data struct {
+	Data string `protobuf:"bytes,2,opt,name=data,proto3,oneof"` // legacy UTF-8 / JSON
+}
+
+type InternalEpistello_BytesData struct {
+	BytesData []byte `protobuf:"bytes,3,opt,name=bytes_data,json=bytesData,proto3,oneof"` // protobuf / binary
+}
+
+func (*InternalEpistello_Data) isInternalEpistello_Payload() {}
+
+func (*InternalEpistello_BytesData) isInternalEpistello_Payload() {}
 
 // Public Epistello message without traceid
 type Epistello struct {
@@ -202,6 +236,69 @@ func (x *Epistello) GetChannel() string {
 	return ""
 }
 
+type EpistelloBytes struct {
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
+	Id      string `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`
+	Data    []byte `protobuf:"bytes,2,opt,name=data,proto3" json:"data,omitempty"`
+	Channel string `protobuf:"bytes,4,opt,name=channel,proto3" json:"channel,omitempty"`
+}
+
+func (x *EpistelloBytes) Reset() {
+	*x = EpistelloBytes{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_proto_eupalinos_proto_msgTypes[2]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *EpistelloBytes) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*EpistelloBytes) ProtoMessage() {}
+
+func (x *EpistelloBytes) ProtoReflect() protoreflect.Message {
+	mi := &file_proto_eupalinos_proto_msgTypes[2]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use EpistelloBytes.ProtoReflect.Descriptor instead.
+func (*EpistelloBytes) Descriptor() ([]byte, []int) {
+	return file_proto_eupalinos_proto_rawDescGZIP(), []int{2}
+}
+
+func (x *EpistelloBytes) GetId() string {
+	if x != nil {
+		return x.Id
+	}
+	return ""
+}
+
+func (x *EpistelloBytes) GetData() []byte {
+	if x != nil {
+		return x.Data
+	}
+	return nil
+}
+
+func (x *EpistelloBytes) GetChannel() string {
+	if x != nil {
+		return x.Channel
+	}
+	return ""
+}
+
 type MessageUpdate struct {
 	state         protoimpl.MessageState
 	sizeCache     protoimpl.SizeCache
@@ -214,7 +311,7 @@ type MessageUpdate struct {
 func (x *MessageUpdate) Reset() {
 	*x = MessageUpdate{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_proto_eupalinos_proto_msgTypes[2]
+		mi := &file_proto_eupalinos_proto_msgTypes[3]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -227,7 +324,7 @@ func (x *MessageUpdate) String() string {
 func (*MessageUpdate) ProtoMessage() {}
 
 func (x *MessageUpdate) ProtoReflect() protoreflect.Message {
-	mi := &file_proto_eupalinos_proto_msgTypes[2]
+	mi := &file_proto_eupalinos_proto_msgTypes[3]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -240,7 +337,7 @@ func (x *MessageUpdate) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use MessageUpdate.ProtoReflect.Descriptor instead.
 func (*MessageUpdate) Descriptor() ([]byte, []int) {
-	return file_proto_eupalinos_proto_rawDescGZIP(), []int{2}
+	return file_proto_eupalinos_proto_rawDescGZIP(), []int{3}
 }
 
 func (x *MessageUpdate) GetOperation() Operation {
@@ -268,7 +365,7 @@ type Diexodos struct {
 func (x *Diexodos) Reset() {
 	*x = Diexodos{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_proto_eupalinos_proto_msgTypes[3]
+		mi := &file_proto_eupalinos_proto_msgTypes[4]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -281,7 +378,7 @@ func (x *Diexodos) String() string {
 func (*Diexodos) ProtoMessage() {}
 
 func (x *Diexodos) ProtoReflect() protoreflect.Message {
-	mi := &file_proto_eupalinos_proto_msgTypes[3]
+	mi := &file_proto_eupalinos_proto_msgTypes[4]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -294,7 +391,7 @@ func (x *Diexodos) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use Diexodos.ProtoReflect.Descriptor instead.
 func (*Diexodos) Descriptor() ([]byte, []int) {
-	return file_proto_eupalinos_proto_rawDescGZIP(), []int{3}
+	return file_proto_eupalinos_proto_rawDescGZIP(), []int{4}
 }
 
 func (x *Diexodos) GetName() string {
@@ -316,7 +413,7 @@ type ChannelInfo struct {
 func (x *ChannelInfo) Reset() {
 	*x = ChannelInfo{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_proto_eupalinos_proto_msgTypes[4]
+		mi := &file_proto_eupalinos_proto_msgTypes[5]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -329,7 +426,7 @@ func (x *ChannelInfo) String() string {
 func (*ChannelInfo) ProtoMessage() {}
 
 func (x *ChannelInfo) ProtoReflect() protoreflect.Message {
-	mi := &file_proto_eupalinos_proto_msgTypes[4]
+	mi := &file_proto_eupalinos_proto_msgTypes[5]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -342,7 +439,7 @@ func (x *ChannelInfo) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use ChannelInfo.ProtoReflect.Descriptor instead.
 func (*ChannelInfo) Descriptor() ([]byte, []int) {
-	return file_proto_eupalinos_proto_rawDescGZIP(), []int{4}
+	return file_proto_eupalinos_proto_rawDescGZIP(), []int{5}
 }
 
 func (x *ChannelInfo) GetName() string {
@@ -361,7 +458,7 @@ type Empty struct {
 func (x *Empty) Reset() {
 	*x = Empty{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_proto_eupalinos_proto_msgTypes[5]
+		mi := &file_proto_eupalinos_proto_msgTypes[6]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -374,7 +471,7 @@ func (x *Empty) String() string {
 func (*Empty) ProtoMessage() {}
 
 func (x *Empty) ProtoReflect() protoreflect.Message {
-	mi := &file_proto_eupalinos_proto_msgTypes[5]
+	mi := &file_proto_eupalinos_proto_msgTypes[6]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -387,7 +484,7 @@ func (x *Empty) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use Empty.ProtoReflect.Descriptor instead.
 func (*Empty) Descriptor() ([]byte, []int) {
-	return file_proto_eupalinos_proto_rawDescGZIP(), []int{5}
+	return file_proto_eupalinos_proto_rawDescGZIP(), []int{6}
 }
 
 // Response message for getting the length of the queue
@@ -402,7 +499,7 @@ type QueueLength struct {
 func (x *QueueLength) Reset() {
 	*x = QueueLength{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_proto_eupalinos_proto_msgTypes[6]
+		mi := &file_proto_eupalinos_proto_msgTypes[7]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -415,7 +512,7 @@ func (x *QueueLength) String() string {
 func (*QueueLength) ProtoMessage() {}
 
 func (x *QueueLength) ProtoReflect() protoreflect.Message {
-	mi := &file_proto_eupalinos_proto_msgTypes[6]
+	mi := &file_proto_eupalinos_proto_msgTypes[7]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -428,7 +525,7 @@ func (x *QueueLength) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use QueueLength.ProtoReflect.Descriptor instead.
 func (*QueueLength) Descriptor() ([]byte, []int) {
-	return file_proto_eupalinos_proto_rawDescGZIP(), []int{6}
+	return file_proto_eupalinos_proto_rawDescGZIP(), []int{7}
 }
 
 func (x *QueueLength) GetLength() int32 {
@@ -449,7 +546,7 @@ type EnqueueResponse struct {
 func (x *EnqueueResponse) Reset() {
 	*x = EnqueueResponse{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_proto_eupalinos_proto_msgTypes[7]
+		mi := &file_proto_eupalinos_proto_msgTypes[8]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -462,7 +559,7 @@ func (x *EnqueueResponse) String() string {
 func (*EnqueueResponse) ProtoMessage() {}
 
 func (x *EnqueueResponse) ProtoReflect() protoreflect.Message {
-	mi := &file_proto_eupalinos_proto_msgTypes[7]
+	mi := &file_proto_eupalinos_proto_msgTypes[8]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -475,7 +572,7 @@ func (x *EnqueueResponse) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use EnqueueResponse.ProtoReflect.Descriptor instead.
 func (*EnqueueResponse) Descriptor() ([]byte, []int) {
-	return file_proto_eupalinos_proto_rawDescGZIP(), []int{7}
+	return file_proto_eupalinos_proto_rawDescGZIP(), []int{8}
 }
 
 func (x *EnqueueResponse) GetId() string {
@@ -485,62 +582,191 @@ func (x *EnqueueResponse) GetId() string {
 	return ""
 }
 
+type HealthRequest struct {
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+}
+
+func (x *HealthRequest) Reset() {
+	*x = HealthRequest{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_proto_eupalinos_proto_msgTypes[9]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *HealthRequest) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*HealthRequest) ProtoMessage() {}
+
+func (x *HealthRequest) ProtoReflect() protoreflect.Message {
+	mi := &file_proto_eupalinos_proto_msgTypes[9]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use HealthRequest.ProtoReflect.Descriptor instead.
+func (*HealthRequest) Descriptor() ([]byte, []int) {
+	return file_proto_eupalinos_proto_rawDescGZIP(), []int{9}
+}
+
+// Response message for health check
+type HealthResponse struct {
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
+	Healthy bool   `protobuf:"varint,1,opt,name=healthy,proto3" json:"healthy,omitempty"`
+	Time    string `protobuf:"bytes,2,opt,name=time,proto3" json:"time,omitempty"`
+	Version string `protobuf:"bytes,3,opt,name=version,proto3" json:"version,omitempty"`
+}
+
+func (x *HealthResponse) Reset() {
+	*x = HealthResponse{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_proto_eupalinos_proto_msgTypes[10]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *HealthResponse) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*HealthResponse) ProtoMessage() {}
+
+func (x *HealthResponse) ProtoReflect() protoreflect.Message {
+	mi := &file_proto_eupalinos_proto_msgTypes[10]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use HealthResponse.ProtoReflect.Descriptor instead.
+func (*HealthResponse) Descriptor() ([]byte, []int) {
+	return file_proto_eupalinos_proto_rawDescGZIP(), []int{10}
+}
+
+func (x *HealthResponse) GetHealthy() bool {
+	if x != nil {
+		return x.Healthy
+	}
+	return false
+}
+
+func (x *HealthResponse) GetTime() string {
+	if x != nil {
+		return x.Time
+	}
+	return ""
+}
+
+func (x *HealthResponse) GetVersion() string {
+	if x != nil {
+		return x.Version
+	}
+	return ""
+}
+
 var File_proto_eupalinos_proto protoreflect.FileDescriptor
 
 var file_proto_eupalinos_proto_rawDesc = []byte{
 	0x0a, 0x15, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2f, 0x65, 0x75, 0x70, 0x61, 0x6c, 0x69, 0x6e, 0x6f,
-	0x73, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12, 0x05, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x22, 0x6b,
-	0x0a, 0x11, 0x49, 0x6e, 0x74, 0x65, 0x72, 0x6e, 0x61, 0x6c, 0x45, 0x70, 0x69, 0x73, 0x74, 0x65,
-	0x6c, 0x6c, 0x6f, 0x12, 0x0e, 0x0a, 0x02, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52,
-	0x02, 0x69, 0x64, 0x12, 0x12, 0x0a, 0x04, 0x64, 0x61, 0x74, 0x61, 0x18, 0x02, 0x20, 0x01, 0x28,
-	0x09, 0x52, 0x04, 0x64, 0x61, 0x74, 0x61, 0x12, 0x18, 0x0a, 0x07, 0x63, 0x68, 0x61, 0x6e, 0x6e,
-	0x65, 0x6c, 0x18, 0x04, 0x20, 0x01, 0x28, 0x09, 0x52, 0x07, 0x63, 0x68, 0x61, 0x6e, 0x6e, 0x65,
-	0x6c, 0x12, 0x18, 0x0a, 0x07, 0x74, 0x72, 0x61, 0x63, 0x65, 0x69, 0x64, 0x18, 0x05, 0x20, 0x01,
-	0x28, 0x09, 0x52, 0x07, 0x74, 0x72, 0x61, 0x63, 0x65, 0x69, 0x64, 0x22, 0x49, 0x0a, 0x09, 0x45,
-	0x70, 0x69, 0x73, 0x74, 0x65, 0x6c, 0x6c, 0x6f, 0x12, 0x0e, 0x0a, 0x02, 0x69, 0x64, 0x18, 0x01,
-	0x20, 0x01, 0x28, 0x09, 0x52, 0x02, 0x69, 0x64, 0x12, 0x12, 0x0a, 0x04, 0x64, 0x61, 0x74, 0x61,
-	0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x04, 0x64, 0x61, 0x74, 0x61, 0x12, 0x18, 0x0a, 0x07,
-	0x63, 0x68, 0x61, 0x6e, 0x6e, 0x65, 0x6c, 0x18, 0x04, 0x20, 0x01, 0x28, 0x09, 0x52, 0x07, 0x63,
-	0x68, 0x61, 0x6e, 0x6e, 0x65, 0x6c, 0x22, 0x73, 0x0a, 0x0d, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67,
-	0x65, 0x55, 0x70, 0x64, 0x61, 0x74, 0x65, 0x12, 0x2e, 0x0a, 0x09, 0x6f, 0x70, 0x65, 0x72, 0x61,
-	0x74, 0x69, 0x6f, 0x6e, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0e, 0x32, 0x10, 0x2e, 0x70, 0x72, 0x6f,
-	0x74, 0x6f, 0x2e, 0x4f, 0x70, 0x65, 0x72, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x52, 0x09, 0x6f, 0x70,
-	0x65, 0x72, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x12, 0x32, 0x0a, 0x07, 0x6d, 0x65, 0x73, 0x73, 0x61,
-	0x67, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x18, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f,
-	0x2e, 0x49, 0x6e, 0x74, 0x65, 0x72, 0x6e, 0x61, 0x6c, 0x45, 0x70, 0x69, 0x73, 0x74, 0x65, 0x6c,
-	0x6c, 0x6f, 0x52, 0x07, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x22, 0x1e, 0x0a, 0x08, 0x44,
-	0x69, 0x65, 0x78, 0x6f, 0x64, 0x6f, 0x73, 0x12, 0x12, 0x0a, 0x04, 0x6e, 0x61, 0x6d, 0x65, 0x18,
-	0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x04, 0x6e, 0x61, 0x6d, 0x65, 0x22, 0x21, 0x0a, 0x0b, 0x43,
-	0x68, 0x61, 0x6e, 0x6e, 0x65, 0x6c, 0x49, 0x6e, 0x66, 0x6f, 0x12, 0x12, 0x0a, 0x04, 0x6e, 0x61,
-	0x6d, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x04, 0x6e, 0x61, 0x6d, 0x65, 0x22, 0x07,
-	0x0a, 0x05, 0x45, 0x6d, 0x70, 0x74, 0x79, 0x22, 0x25, 0x0a, 0x0b, 0x51, 0x75, 0x65, 0x75, 0x65,
-	0x4c, 0x65, 0x6e, 0x67, 0x74, 0x68, 0x12, 0x16, 0x0a, 0x06, 0x6c, 0x65, 0x6e, 0x67, 0x74, 0x68,
-	0x18, 0x01, 0x20, 0x01, 0x28, 0x05, 0x52, 0x06, 0x6c, 0x65, 0x6e, 0x67, 0x74, 0x68, 0x22, 0x21,
-	0x0a, 0x0f, 0x45, 0x6e, 0x71, 0x75, 0x65, 0x75, 0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73,
-	0x65, 0x12, 0x0e, 0x0a, 0x02, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x02, 0x69,
-	0x64, 0x2a, 0x25, 0x0a, 0x09, 0x4f, 0x70, 0x65, 0x72, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x12, 0x0b,
-	0x0a, 0x07, 0x45, 0x4e, 0x51, 0x55, 0x45, 0x55, 0x45, 0x10, 0x00, 0x12, 0x0b, 0x0a, 0x07, 0x44,
-	0x45, 0x51, 0x55, 0x45, 0x55, 0x45, 0x10, 0x01, 0x32, 0x83, 0x02, 0x0a, 0x09, 0x45, 0x75, 0x70,
-	0x61, 0x6c, 0x69, 0x6e, 0x6f, 0x73, 0x12, 0x48, 0x0a, 0x12, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d,
-	0x51, 0x75, 0x65, 0x75, 0x65, 0x55, 0x70, 0x64, 0x61, 0x74, 0x65, 0x73, 0x12, 0x14, 0x2e, 0x70,
-	0x72, 0x6f, 0x74, 0x6f, 0x2e, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x55, 0x70, 0x64, 0x61,
-	0x74, 0x65, 0x1a, 0x18, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2e, 0x49, 0x6e, 0x74, 0x65, 0x72,
-	0x6e, 0x61, 0x6c, 0x45, 0x70, 0x69, 0x73, 0x74, 0x65, 0x6c, 0x6c, 0x6f, 0x28, 0x01, 0x30, 0x01,
-	0x12, 0x3a, 0x0a, 0x0e, 0x45, 0x6e, 0x71, 0x75, 0x65, 0x75, 0x65, 0x4d, 0x65, 0x73, 0x73, 0x61,
-	0x67, 0x65, 0x12, 0x10, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2e, 0x45, 0x70, 0x69, 0x73, 0x74,
-	0x65, 0x6c, 0x6c, 0x6f, 0x1a, 0x16, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2e, 0x45, 0x6e, 0x71,
-	0x75, 0x65, 0x75, 0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x36, 0x0a, 0x0e,
-	0x44, 0x65, 0x71, 0x75, 0x65, 0x75, 0x65, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x12, 0x12,
-	0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2e, 0x43, 0x68, 0x61, 0x6e, 0x6e, 0x65, 0x6c, 0x49, 0x6e,
-	0x66, 0x6f, 0x1a, 0x10, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2e, 0x45, 0x70, 0x69, 0x73, 0x74,
-	0x65, 0x6c, 0x6c, 0x6f, 0x12, 0x38, 0x0a, 0x0e, 0x47, 0x65, 0x74, 0x51, 0x75, 0x65, 0x75, 0x65,
-	0x4c, 0x65, 0x6e, 0x67, 0x74, 0x68, 0x12, 0x12, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2e, 0x43,
-	0x68, 0x61, 0x6e, 0x6e, 0x65, 0x6c, 0x49, 0x6e, 0x66, 0x6f, 0x1a, 0x12, 0x2e, 0x70, 0x72, 0x6f,
-	0x74, 0x6f, 0x2e, 0x51, 0x75, 0x65, 0x75, 0x65, 0x4c, 0x65, 0x6e, 0x67, 0x74, 0x68, 0x42, 0x2b,
-	0x5a, 0x29, 0x67, 0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d, 0x2f, 0x6f, 0x64, 0x79,
-	0x73, 0x73, 0x65, 0x69, 0x61, 0x2d, 0x67, 0x72, 0x65, 0x65, 0x6b, 0x2f, 0x65, 0x75, 0x70, 0x61,
-	0x6c, 0x69, 0x6e, 0x6f, 0x73, 0x2f, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x06, 0x70, 0x72, 0x6f,
-	0x74, 0x6f, 0x33,
+	0x73, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12, 0x05, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x22, 0x99,
+	0x01, 0x0a, 0x11, 0x49, 0x6e, 0x74, 0x65, 0x72, 0x6e, 0x61, 0x6c, 0x45, 0x70, 0x69, 0x73, 0x74,
+	0x65, 0x6c, 0x6c, 0x6f, 0x12, 0x0e, 0x0a, 0x02, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09,
+	0x52, 0x02, 0x69, 0x64, 0x12, 0x14, 0x0a, 0x04, 0x64, 0x61, 0x74, 0x61, 0x18, 0x02, 0x20, 0x01,
+	0x28, 0x09, 0x48, 0x00, 0x52, 0x04, 0x64, 0x61, 0x74, 0x61, 0x12, 0x1f, 0x0a, 0x0a, 0x62, 0x79,
+	0x74, 0x65, 0x73, 0x5f, 0x64, 0x61, 0x74, 0x61, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0c, 0x48, 0x00,
+	0x52, 0x09, 0x62, 0x79, 0x74, 0x65, 0x73, 0x44, 0x61, 0x74, 0x61, 0x12, 0x18, 0x0a, 0x07, 0x63,
+	0x68, 0x61, 0x6e, 0x6e, 0x65, 0x6c, 0x18, 0x04, 0x20, 0x01, 0x28, 0x09, 0x52, 0x07, 0x63, 0x68,
+	0x61, 0x6e, 0x6e, 0x65, 0x6c, 0x12, 0x18, 0x0a, 0x07, 0x74, 0x72, 0x61, 0x63, 0x65, 0x69, 0x64,
+	0x18, 0x05, 0x20, 0x01, 0x28, 0x09, 0x52, 0x07, 0x74, 0x72, 0x61, 0x63, 0x65, 0x69, 0x64, 0x42,
+	0x09, 0x0a, 0x07, 0x70, 0x61, 0x79, 0x6c, 0x6f, 0x61, 0x64, 0x22, 0x49, 0x0a, 0x09, 0x45, 0x70,
+	0x69, 0x73, 0x74, 0x65, 0x6c, 0x6c, 0x6f, 0x12, 0x0e, 0x0a, 0x02, 0x69, 0x64, 0x18, 0x01, 0x20,
+	0x01, 0x28, 0x09, 0x52, 0x02, 0x69, 0x64, 0x12, 0x12, 0x0a, 0x04, 0x64, 0x61, 0x74, 0x61, 0x18,
+	0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x04, 0x64, 0x61, 0x74, 0x61, 0x12, 0x18, 0x0a, 0x07, 0x63,
+	0x68, 0x61, 0x6e, 0x6e, 0x65, 0x6c, 0x18, 0x04, 0x20, 0x01, 0x28, 0x09, 0x52, 0x07, 0x63, 0x68,
+	0x61, 0x6e, 0x6e, 0x65, 0x6c, 0x22, 0x4e, 0x0a, 0x0e, 0x45, 0x70, 0x69, 0x73, 0x74, 0x65, 0x6c,
+	0x6c, 0x6f, 0x42, 0x79, 0x74, 0x65, 0x73, 0x12, 0x0e, 0x0a, 0x02, 0x69, 0x64, 0x18, 0x01, 0x20,
+	0x01, 0x28, 0x09, 0x52, 0x02, 0x69, 0x64, 0x12, 0x12, 0x0a, 0x04, 0x64, 0x61, 0x74, 0x61, 0x18,
+	0x02, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x04, 0x64, 0x61, 0x74, 0x61, 0x12, 0x18, 0x0a, 0x07, 0x63,
+	0x68, 0x61, 0x6e, 0x6e, 0x65, 0x6c, 0x18, 0x04, 0x20, 0x01, 0x28, 0x09, 0x52, 0x07, 0x63, 0x68,
+	0x61, 0x6e, 0x6e, 0x65, 0x6c, 0x22, 0x73, 0x0a, 0x0d, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65,
+	0x55, 0x70, 0x64, 0x61, 0x74, 0x65, 0x12, 0x2e, 0x0a, 0x09, 0x6f, 0x70, 0x65, 0x72, 0x61, 0x74,
+	0x69, 0x6f, 0x6e, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0e, 0x32, 0x10, 0x2e, 0x70, 0x72, 0x6f, 0x74,
+	0x6f, 0x2e, 0x4f, 0x70, 0x65, 0x72, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x52, 0x09, 0x6f, 0x70, 0x65,
+	0x72, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x12, 0x32, 0x0a, 0x07, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67,
+	0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x18, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2e,
+	0x49, 0x6e, 0x74, 0x65, 0x72, 0x6e, 0x61, 0x6c, 0x45, 0x70, 0x69, 0x73, 0x74, 0x65, 0x6c, 0x6c,
+	0x6f, 0x52, 0x07, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x22, 0x1e, 0x0a, 0x08, 0x44, 0x69,
+	0x65, 0x78, 0x6f, 0x64, 0x6f, 0x73, 0x12, 0x12, 0x0a, 0x04, 0x6e, 0x61, 0x6d, 0x65, 0x18, 0x01,
+	0x20, 0x01, 0x28, 0x09, 0x52, 0x04, 0x6e, 0x61, 0x6d, 0x65, 0x22, 0x21, 0x0a, 0x0b, 0x43, 0x68,
+	0x61, 0x6e, 0x6e, 0x65, 0x6c, 0x49, 0x6e, 0x66, 0x6f, 0x12, 0x12, 0x0a, 0x04, 0x6e, 0x61, 0x6d,
+	0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x04, 0x6e, 0x61, 0x6d, 0x65, 0x22, 0x07, 0x0a,
+	0x05, 0x45, 0x6d, 0x70, 0x74, 0x79, 0x22, 0x25, 0x0a, 0x0b, 0x51, 0x75, 0x65, 0x75, 0x65, 0x4c,
+	0x65, 0x6e, 0x67, 0x74, 0x68, 0x12, 0x16, 0x0a, 0x06, 0x6c, 0x65, 0x6e, 0x67, 0x74, 0x68, 0x18,
+	0x01, 0x20, 0x01, 0x28, 0x05, 0x52, 0x06, 0x6c, 0x65, 0x6e, 0x67, 0x74, 0x68, 0x22, 0x21, 0x0a,
+	0x0f, 0x45, 0x6e, 0x71, 0x75, 0x65, 0x75, 0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65,
+	0x12, 0x0e, 0x0a, 0x02, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x02, 0x69, 0x64,
+	0x22, 0x0f, 0x0a, 0x0d, 0x48, 0x65, 0x61, 0x6c, 0x74, 0x68, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73,
+	0x74, 0x22, 0x58, 0x0a, 0x0e, 0x48, 0x65, 0x61, 0x6c, 0x74, 0x68, 0x52, 0x65, 0x73, 0x70, 0x6f,
+	0x6e, 0x73, 0x65, 0x12, 0x18, 0x0a, 0x07, 0x68, 0x65, 0x61, 0x6c, 0x74, 0x68, 0x79, 0x18, 0x01,
+	0x20, 0x01, 0x28, 0x08, 0x52, 0x07, 0x68, 0x65, 0x61, 0x6c, 0x74, 0x68, 0x79, 0x12, 0x12, 0x0a,
+	0x04, 0x74, 0x69, 0x6d, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x04, 0x74, 0x69, 0x6d,
+	0x65, 0x12, 0x18, 0x0a, 0x07, 0x76, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x18, 0x03, 0x20, 0x01,
+	0x28, 0x09, 0x52, 0x07, 0x76, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x2a, 0x25, 0x0a, 0x09, 0x4f,
+	0x70, 0x65, 0x72, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x12, 0x0b, 0x0a, 0x07, 0x45, 0x4e, 0x51, 0x55,
+	0x45, 0x55, 0x45, 0x10, 0x00, 0x12, 0x0b, 0x0a, 0x07, 0x44, 0x45, 0x51, 0x55, 0x45, 0x55, 0x45,
+	0x10, 0x01, 0x32, 0xc4, 0x03, 0x0a, 0x09, 0x45, 0x75, 0x70, 0x61, 0x6c, 0x69, 0x6e, 0x6f, 0x73,
+	0x12, 0x37, 0x0a, 0x06, 0x48, 0x65, 0x61, 0x6c, 0x74, 0x68, 0x12, 0x14, 0x2e, 0x70, 0x72, 0x6f,
+	0x74, 0x6f, 0x2e, 0x48, 0x65, 0x61, 0x6c, 0x74, 0x68, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74,
+	0x1a, 0x15, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2e, 0x48, 0x65, 0x61, 0x6c, 0x74, 0x68, 0x52,
+	0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x00, 0x12, 0x48, 0x0a, 0x12, 0x53, 0x74, 0x72,
+	0x65, 0x61, 0x6d, 0x51, 0x75, 0x65, 0x75, 0x65, 0x55, 0x70, 0x64, 0x61, 0x74, 0x65, 0x73, 0x12,
+	0x14, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2e, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x55,
+	0x70, 0x64, 0x61, 0x74, 0x65, 0x1a, 0x18, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2e, 0x49, 0x6e,
+	0x74, 0x65, 0x72, 0x6e, 0x61, 0x6c, 0x45, 0x70, 0x69, 0x73, 0x74, 0x65, 0x6c, 0x6c, 0x6f, 0x28,
+	0x01, 0x30, 0x01, 0x12, 0x3a, 0x0a, 0x0e, 0x45, 0x6e, 0x71, 0x75, 0x65, 0x75, 0x65, 0x4d, 0x65,
+	0x73, 0x73, 0x61, 0x67, 0x65, 0x12, 0x10, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2e, 0x45, 0x70,
+	0x69, 0x73, 0x74, 0x65, 0x6c, 0x6c, 0x6f, 0x1a, 0x16, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2e,
+	0x45, 0x6e, 0x71, 0x75, 0x65, 0x75, 0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12,
+	0x44, 0x0a, 0x13, 0x45, 0x6e, 0x71, 0x75, 0x65, 0x75, 0x65, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67,
+	0x65, 0x42, 0x79, 0x74, 0x65, 0x73, 0x12, 0x15, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2e, 0x45,
+	0x70, 0x69, 0x73, 0x74, 0x65, 0x6c, 0x6c, 0x6f, 0x42, 0x79, 0x74, 0x65, 0x73, 0x1a, 0x16, 0x2e,
+	0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2e, 0x45, 0x6e, 0x71, 0x75, 0x65, 0x75, 0x65, 0x52, 0x65, 0x73,
+	0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x36, 0x0a, 0x0e, 0x44, 0x65, 0x71, 0x75, 0x65, 0x75, 0x65,
+	0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x12, 0x12, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2e,
+	0x43, 0x68, 0x61, 0x6e, 0x6e, 0x65, 0x6c, 0x49, 0x6e, 0x66, 0x6f, 0x1a, 0x10, 0x2e, 0x70, 0x72,
+	0x6f, 0x74, 0x6f, 0x2e, 0x45, 0x70, 0x69, 0x73, 0x74, 0x65, 0x6c, 0x6c, 0x6f, 0x12, 0x40, 0x0a,
+	0x13, 0x44, 0x65, 0x71, 0x75, 0x65, 0x75, 0x65, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x42,
+	0x79, 0x74, 0x65, 0x73, 0x12, 0x12, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2e, 0x43, 0x68, 0x61,
+	0x6e, 0x6e, 0x65, 0x6c, 0x49, 0x6e, 0x66, 0x6f, 0x1a, 0x15, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f,
+	0x2e, 0x45, 0x70, 0x69, 0x73, 0x74, 0x65, 0x6c, 0x6c, 0x6f, 0x42, 0x79, 0x74, 0x65, 0x73, 0x12,
+	0x38, 0x0a, 0x0e, 0x47, 0x65, 0x74, 0x51, 0x75, 0x65, 0x75, 0x65, 0x4c, 0x65, 0x6e, 0x67, 0x74,
+	0x68, 0x12, 0x12, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2e, 0x43, 0x68, 0x61, 0x6e, 0x6e, 0x65,
+	0x6c, 0x49, 0x6e, 0x66, 0x6f, 0x1a, 0x12, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2e, 0x51, 0x75,
+	0x65, 0x75, 0x65, 0x4c, 0x65, 0x6e, 0x67, 0x74, 0x68, 0x42, 0x31, 0x5a, 0x2f, 0x67, 0x69, 0x74,
+	0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d, 0x2f, 0x6f, 0x64, 0x79, 0x73, 0x73, 0x65, 0x69, 0x61,
+	0x2d, 0x67, 0x72, 0x65, 0x65, 0x6b, 0x2f, 0x61, 0x67, 0x6f, 0x72, 0x61, 0x2f, 0x65, 0x75, 0x70,
+	0x61, 0x6c, 0x69, 0x6e, 0x6f, 0x73, 0x2f, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x06, 0x70, 0x72,
+	0x6f, 0x74, 0x6f, 0x33,
 }
 
 var (
@@ -556,34 +782,43 @@ func file_proto_eupalinos_proto_rawDescGZIP() []byte {
 }
 
 var file_proto_eupalinos_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
-var file_proto_eupalinos_proto_msgTypes = make([]protoimpl.MessageInfo, 8)
+var file_proto_eupalinos_proto_msgTypes = make([]protoimpl.MessageInfo, 11)
 var file_proto_eupalinos_proto_goTypes = []interface{}{
 	(Operation)(0),            // 0: proto.Operation
 	(*InternalEpistello)(nil), // 1: proto.InternalEpistello
 	(*Epistello)(nil),         // 2: proto.Epistello
-	(*MessageUpdate)(nil),     // 3: proto.MessageUpdate
-	(*Diexodos)(nil),          // 4: proto.Diexodos
-	(*ChannelInfo)(nil),       // 5: proto.ChannelInfo
-	(*Empty)(nil),             // 6: proto.Empty
-	(*QueueLength)(nil),       // 7: proto.QueueLength
-	(*EnqueueResponse)(nil),   // 8: proto.EnqueueResponse
+	(*EpistelloBytes)(nil),    // 3: proto.EpistelloBytes
+	(*MessageUpdate)(nil),     // 4: proto.MessageUpdate
+	(*Diexodos)(nil),          // 5: proto.Diexodos
+	(*ChannelInfo)(nil),       // 6: proto.ChannelInfo
+	(*Empty)(nil),             // 7: proto.Empty
+	(*QueueLength)(nil),       // 8: proto.QueueLength
+	(*EnqueueResponse)(nil),   // 9: proto.EnqueueResponse
+	(*HealthRequest)(nil),     // 10: proto.HealthRequest
+	(*HealthResponse)(nil),    // 11: proto.HealthResponse
 }
 var file_proto_eupalinos_proto_depIdxs = []int32{
-	0, // 0: proto.MessageUpdate.operation:type_name -> proto.Operation
-	1, // 1: proto.MessageUpdate.message:type_name -> proto.InternalEpistello
-	3, // 2: proto.Eupalinos.StreamQueueUpdates:input_type -> proto.MessageUpdate
-	2, // 3: proto.Eupalinos.EnqueueMessage:input_type -> proto.Epistello
-	5, // 4: proto.Eupalinos.DequeueMessage:input_type -> proto.ChannelInfo
-	5, // 5: proto.Eupalinos.GetQueueLength:input_type -> proto.ChannelInfo
-	1, // 6: proto.Eupalinos.StreamQueueUpdates:output_type -> proto.InternalEpistello
-	8, // 7: proto.Eupalinos.EnqueueMessage:output_type -> proto.EnqueueResponse
-	2, // 8: proto.Eupalinos.DequeueMessage:output_type -> proto.Epistello
-	7, // 9: proto.Eupalinos.GetQueueLength:output_type -> proto.QueueLength
-	6, // [6:10] is the sub-list for method output_type
-	2, // [2:6] is the sub-list for method input_type
-	2, // [2:2] is the sub-list for extension type_name
-	2, // [2:2] is the sub-list for extension extendee
-	0, // [0:2] is the sub-list for field type_name
+	0,  // 0: proto.MessageUpdate.operation:type_name -> proto.Operation
+	1,  // 1: proto.MessageUpdate.message:type_name -> proto.InternalEpistello
+	10, // 2: proto.Eupalinos.Health:input_type -> proto.HealthRequest
+	4,  // 3: proto.Eupalinos.StreamQueueUpdates:input_type -> proto.MessageUpdate
+	2,  // 4: proto.Eupalinos.EnqueueMessage:input_type -> proto.Epistello
+	3,  // 5: proto.Eupalinos.EnqueueMessageBytes:input_type -> proto.EpistelloBytes
+	6,  // 6: proto.Eupalinos.DequeueMessage:input_type -> proto.ChannelInfo
+	6,  // 7: proto.Eupalinos.DequeueMessageBytes:input_type -> proto.ChannelInfo
+	6,  // 8: proto.Eupalinos.GetQueueLength:input_type -> proto.ChannelInfo
+	11, // 9: proto.Eupalinos.Health:output_type -> proto.HealthResponse
+	1,  // 10: proto.Eupalinos.StreamQueueUpdates:output_type -> proto.InternalEpistello
+	9,  // 11: proto.Eupalinos.EnqueueMessage:output_type -> proto.EnqueueResponse
+	9,  // 12: proto.Eupalinos.EnqueueMessageBytes:output_type -> proto.EnqueueResponse
+	2,  // 13: proto.Eupalinos.DequeueMessage:output_type -> proto.Epistello
+	3,  // 14: proto.Eupalinos.DequeueMessageBytes:output_type -> proto.EpistelloBytes
+	8,  // 15: proto.Eupalinos.GetQueueLength:output_type -> proto.QueueLength
+	9,  // [9:16] is the sub-list for method output_type
+	2,  // [2:9] is the sub-list for method input_type
+	2,  // [2:2] is the sub-list for extension type_name
+	2,  // [2:2] is the sub-list for extension extendee
+	0,  // [0:2] is the sub-list for field type_name
 }
 
 func init() { file_proto_eupalinos_proto_init() }
@@ -617,7 +852,7 @@ func file_proto_eupalinos_proto_init() {
 			}
 		}
 		file_proto_eupalinos_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*MessageUpdate); i {
+			switch v := v.(*EpistelloBytes); i {
 			case 0:
 				return &v.state
 			case 1:
@@ -629,7 +864,7 @@ func file_proto_eupalinos_proto_init() {
 			}
 		}
 		file_proto_eupalinos_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Diexodos); i {
+			switch v := v.(*MessageUpdate); i {
 			case 0:
 				return &v.state
 			case 1:
@@ -641,7 +876,7 @@ func file_proto_eupalinos_proto_init() {
 			}
 		}
 		file_proto_eupalinos_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*ChannelInfo); i {
+			switch v := v.(*Diexodos); i {
 			case 0:
 				return &v.state
 			case 1:
@@ -653,7 +888,7 @@ func file_proto_eupalinos_proto_init() {
 			}
 		}
 		file_proto_eupalinos_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Empty); i {
+			switch v := v.(*ChannelInfo); i {
 			case 0:
 				return &v.state
 			case 1:
@@ -665,7 +900,7 @@ func file_proto_eupalinos_proto_init() {
 			}
 		}
 		file_proto_eupalinos_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*QueueLength); i {
+			switch v := v.(*Empty); i {
 			case 0:
 				return &v.state
 			case 1:
@@ -677,6 +912,18 @@ func file_proto_eupalinos_proto_init() {
 			}
 		}
 		file_proto_eupalinos_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
+			switch v := v.(*QueueLength); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
+		file_proto_eupalinos_proto_msgTypes[8].Exporter = func(v interface{}, i int) interface{} {
 			switch v := v.(*EnqueueResponse); i {
 			case 0:
 				return &v.state
@@ -688,6 +935,34 @@ func file_proto_eupalinos_proto_init() {
 				return nil
 			}
 		}
+		file_proto_eupalinos_proto_msgTypes[9].Exporter = func(v interface{}, i int) interface{} {
+			switch v := v.(*HealthRequest); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
+		file_proto_eupalinos_proto_msgTypes[10].Exporter = func(v interface{}, i int) interface{} {
+			switch v := v.(*HealthResponse); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
+	}
+	file_proto_eupalinos_proto_msgTypes[0].OneofWrappers = []interface{}{
+		(*InternalEpistello_Data)(nil),
+		(*InternalEpistello_BytesData)(nil),
 	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
@@ -695,7 +970,7 @@ func file_proto_eupalinos_proto_init() {
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
 			RawDescriptor: file_proto_eupalinos_proto_rawDesc,
 			NumEnums:      1,
-			NumMessages:   8,
+			NumMessages:   11,
 			NumExtensions: 0,
 			NumServices:   1,
 		},
