@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
 // - protoc-gen-go-grpc v1.2.0
-// - protoc             v3.21.12
+// - protoc             v6.33.2
 // source: proto/eupalinos.proto
 
 package proto
@@ -22,12 +22,15 @@ const _ = grpc.SupportPackageIsVersion7
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
 type EupalinosClient interface {
+	Health(ctx context.Context, in *HealthRequest, opts ...grpc.CallOption) (*HealthResponse, error)
 	// Bidirectional Streaming for task updates between Eupalinos pods
 	StreamQueueUpdates(ctx context.Context, opts ...grpc.CallOption) (Eupalinos_StreamQueueUpdatesClient, error)
 	// Unary RPC for epistello enqueueing
 	EnqueueMessage(ctx context.Context, in *Epistello, opts ...grpc.CallOption) (*EnqueueResponse, error)
+	EnqueueMessageBytes(ctx context.Context, in *EpistelloBytes, opts ...grpc.CallOption) (*EnqueueResponse, error)
 	// Unary RPC for epistello dequeueing
 	DequeueMessage(ctx context.Context, in *ChannelInfo, opts ...grpc.CallOption) (*Epistello, error)
+	DequeueMessageBytes(ctx context.Context, in *ChannelInfo, opts ...grpc.CallOption) (*EpistelloBytes, error)
 	// Unary RPC for getting the length of the queue
 	GetQueueLength(ctx context.Context, in *ChannelInfo, opts ...grpc.CallOption) (*QueueLength, error)
 }
@@ -38,6 +41,15 @@ type eupalinosClient struct {
 
 func NewEupalinosClient(cc grpc.ClientConnInterface) EupalinosClient {
 	return &eupalinosClient{cc}
+}
+
+func (c *eupalinosClient) Health(ctx context.Context, in *HealthRequest, opts ...grpc.CallOption) (*HealthResponse, error) {
+	out := new(HealthResponse)
+	err := c.cc.Invoke(ctx, "/proto.Eupalinos/Health", in, out, opts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
 }
 
 func (c *eupalinosClient) StreamQueueUpdates(ctx context.Context, opts ...grpc.CallOption) (Eupalinos_StreamQueueUpdatesClient, error) {
@@ -80,9 +92,27 @@ func (c *eupalinosClient) EnqueueMessage(ctx context.Context, in *Epistello, opt
 	return out, nil
 }
 
+func (c *eupalinosClient) EnqueueMessageBytes(ctx context.Context, in *EpistelloBytes, opts ...grpc.CallOption) (*EnqueueResponse, error) {
+	out := new(EnqueueResponse)
+	err := c.cc.Invoke(ctx, "/proto.Eupalinos/EnqueueMessageBytes", in, out, opts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
 func (c *eupalinosClient) DequeueMessage(ctx context.Context, in *ChannelInfo, opts ...grpc.CallOption) (*Epistello, error) {
 	out := new(Epistello)
 	err := c.cc.Invoke(ctx, "/proto.Eupalinos/DequeueMessage", in, out, opts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
+func (c *eupalinosClient) DequeueMessageBytes(ctx context.Context, in *ChannelInfo, opts ...grpc.CallOption) (*EpistelloBytes, error) {
+	out := new(EpistelloBytes)
+	err := c.cc.Invoke(ctx, "/proto.Eupalinos/DequeueMessageBytes", in, out, opts...)
 	if err != nil {
 		return nil, err
 	}
@@ -102,12 +132,15 @@ func (c *eupalinosClient) GetQueueLength(ctx context.Context, in *ChannelInfo, o
 // All implementations must embed UnimplementedEupalinosServer
 // for forward compatibility
 type EupalinosServer interface {
+	Health(context.Context, *HealthRequest) (*HealthResponse, error)
 	// Bidirectional Streaming for task updates between Eupalinos pods
 	StreamQueueUpdates(Eupalinos_StreamQueueUpdatesServer) error
 	// Unary RPC for epistello enqueueing
 	EnqueueMessage(context.Context, *Epistello) (*EnqueueResponse, error)
+	EnqueueMessageBytes(context.Context, *EpistelloBytes) (*EnqueueResponse, error)
 	// Unary RPC for epistello dequeueing
 	DequeueMessage(context.Context, *ChannelInfo) (*Epistello, error)
+	DequeueMessageBytes(context.Context, *ChannelInfo) (*EpistelloBytes, error)
 	// Unary RPC for getting the length of the queue
 	GetQueueLength(context.Context, *ChannelInfo) (*QueueLength, error)
 	mustEmbedUnimplementedEupalinosServer()
@@ -117,14 +150,23 @@ type EupalinosServer interface {
 type UnimplementedEupalinosServer struct {
 }
 
+func (UnimplementedEupalinosServer) Health(context.Context, *HealthRequest) (*HealthResponse, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method Health not implemented")
+}
 func (UnimplementedEupalinosServer) StreamQueueUpdates(Eupalinos_StreamQueueUpdatesServer) error {
 	return status.Errorf(codes.Unimplemented, "method StreamQueueUpdates not implemented")
 }
 func (UnimplementedEupalinosServer) EnqueueMessage(context.Context, *Epistello) (*EnqueueResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method EnqueueMessage not implemented")
 }
+func (UnimplementedEupalinosServer) EnqueueMessageBytes(context.Context, *EpistelloBytes) (*EnqueueResponse, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method EnqueueMessageBytes not implemented")
+}
 func (UnimplementedEupalinosServer) DequeueMessage(context.Context, *ChannelInfo) (*Epistello, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method DequeueMessage not implemented")
+}
+func (UnimplementedEupalinosServer) DequeueMessageBytes(context.Context, *ChannelInfo) (*EpistelloBytes, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method DequeueMessageBytes not implemented")
 }
 func (UnimplementedEupalinosServer) GetQueueLength(context.Context, *ChannelInfo) (*QueueLength, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetQueueLength not implemented")
@@ -140,6 +182,24 @@ type UnsafeEupalinosServer interface {
 
 func RegisterEupalinosServer(s grpc.ServiceRegistrar, srv EupalinosServer) {
 	s.RegisterService(&Eupalinos_ServiceDesc, srv)
+}
+
+func _Eupalinos_Health_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(HealthRequest)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(EupalinosServer).Health(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: "/proto.Eupalinos/Health",
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(EupalinosServer).Health(ctx, req.(*HealthRequest))
+	}
+	return interceptor(ctx, in, info, handler)
 }
 
 func _Eupalinos_StreamQueueUpdates_Handler(srv interface{}, stream grpc.ServerStream) error {
@@ -186,6 +246,24 @@ func _Eupalinos_EnqueueMessage_Handler(srv interface{}, ctx context.Context, dec
 	return interceptor(ctx, in, info, handler)
 }
 
+func _Eupalinos_EnqueueMessageBytes_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(EpistelloBytes)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(EupalinosServer).EnqueueMessageBytes(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: "/proto.Eupalinos/EnqueueMessageBytes",
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(EupalinosServer).EnqueueMessageBytes(ctx, req.(*EpistelloBytes))
+	}
+	return interceptor(ctx, in, info, handler)
+}
+
 func _Eupalinos_DequeueMessage_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
 	in := new(ChannelInfo)
 	if err := dec(in); err != nil {
@@ -200,6 +278,24 @@ func _Eupalinos_DequeueMessage_Handler(srv interface{}, ctx context.Context, dec
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(EupalinosServer).DequeueMessage(ctx, req.(*ChannelInfo))
+	}
+	return interceptor(ctx, in, info, handler)
+}
+
+func _Eupalinos_DequeueMessageBytes_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(ChannelInfo)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(EupalinosServer).DequeueMessageBytes(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: "/proto.Eupalinos/DequeueMessageBytes",
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(EupalinosServer).DequeueMessageBytes(ctx, req.(*ChannelInfo))
 	}
 	return interceptor(ctx, in, info, handler)
 }
@@ -230,12 +326,24 @@ var Eupalinos_ServiceDesc = grpc.ServiceDesc{
 	HandlerType: (*EupalinosServer)(nil),
 	Methods: []grpc.MethodDesc{
 		{
+			MethodName: "Health",
+			Handler:    _Eupalinos_Health_Handler,
+		},
+		{
 			MethodName: "EnqueueMessage",
 			Handler:    _Eupalinos_EnqueueMessage_Handler,
 		},
 		{
+			MethodName: "EnqueueMessageBytes",
+			Handler:    _Eupalinos_EnqueueMessageBytes_Handler,
+		},
+		{
 			MethodName: "DequeueMessage",
 			Handler:    _Eupalinos_DequeueMessage_Handler,
+		},
+		{
+			MethodName: "DequeueMessageBytes",
+			Handler:    _Eupalinos_DequeueMessageBytes_Handler,
 		},
 		{
 			MethodName: "GetQueueLength",
